Loading add field attributes from schema

+++ Core/Fields/Float64Field.cs
@@ -3,8 +3,8 @@ namespace Sharara.EntityCodeGen.Core.Fields
     class Float64Field : Field
     {
         public const string XmlTypeName = "f64";
-        public long MinValue { get; set; } = long.MinValue;
-        public long MaxValue { get; set; } = long.MaxValue;
+        public double MinValue { get; set; } = double.MinValue;
+        public double MaxValue { get; set; } = double.MaxValue;
 
         public Float64Field(string name)
             : base(FieldType.Float64.Instance, name)
